requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
numpy==1.26.0
pandas==2.2.0
yfinance==0.2.36
scikit-learn==0.24.2
matplotlib==3.8.0
cvxpy==1.1.15
statsmodels==0.13.0
seaborn==0.11.2
plotly==5.3.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
email-validator==1.1.3
web3==5.24.0
Flask-RESTful==0.3.10
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
python-jose==3.3.0
requests==2.31.0
streamlit==1.32.0
scipy==1.12.0 